ctypes.pythonapi isn't defined on pypy, so don't require it for testing.
We don't currently implement pygame's array interface, so the testing
helpers in test/test_utils/arrinter.py aren't currently useful.
However, they are unconditionaly imported, which causes a bunch
of spurious failures when running the test suite on pypy, due
to arrintr.py throwing NameError because of the non-existant module.

This avoids those failure by ensuring we create stub functions for
this case. If we ever implement the array interface, these helpers
will need to be complete rewritten.