Fix crash in smoothscale
As reported on irc.

Smoothscale handled the non-transform case incorrectly when the
pitch of the destination surface was not the same as the the original
surface (which happens when dealing with subsurfaces as the source,
for example). This changes the logic to copy the image a row at a
time to avoid that issue (which matches pygame's logic).