Support more surface depths in the conformance test suite
We add support for running the conformance tests on additional
surface depths.

To simplify specifying the depths for test cases, we add a decorator.
We also use this to specify the test name, which will somewhat
simplify adding additional conformance tests in the future.

We tweak a few test cases to be sure we are exercising the colour
conversion path when testing 8-bit surfaces.

This also adds an exit code and verbose flag to the test runner,
with an eye to running the conformance test on our CI infrastructure.